Update json-ui-intro.md
This adds the button mappings section, and some detailed modifications and changes to other parts. Mainly adjusting some texts to make it as simpler as possible. The purpose of this is to make the intro page where the only basics of JSON UI is outlined for newcomers - and at the same time the experienced ones too.