Updated attachables article
A lot of the phrasing has been improved to be more clear.
Updated the instructions and attached ZIP file to be easier to use; instead of requiring users to move their cubes and bones to the arm, an offset was implemented in the animation to do this automatically. Users still need to set the arm as the parent however. The left arm was also added if players wanted to utilize that too.
Made it more clear that the format version for the geometry should be "1.16.0" to use model binding.
Updated images to have consistent 16:9 resolutions.