Update format-version.md to add info that for some items/blocks you may need a format version higher than 1.16.100.
Added info that for some items/blocks you may need a format version higher than 1.16.100. One example of this is that to make a custom spawn egg that you made able to set the entity type of a monster spawner then the format version needs to be set to at least 1.19.80 or higher.